+++ src/ethereum/fees/feeProviders.ts
@@ -210,18 +210,7 @@ export const fetchFeesFromEvmGasStation = async (
   const result = await fetch(`${evmGasStationUrl}${apiKeyParams}`)
   const jsonObj = await result.json()
 
-  const fees = asEvmGasStation(jsonObj)
-  // Special case for POL fast and fastest being equivalent from gas station
-  if (currencyInfo.currencyCode === 'POL') {
-    // Since the later code assumes EthGasStation's
-    // greater-by-a-factor-of-ten gas prices, we need to multiply the GWEI
-    // from Polygon Gas Station by 10 so they conform.
-    fees.safeLow *= 10
-    // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
-    fees.average = ((jsonObj.fast + jsonObj.safeLow) / 2) * 10
-    fees.fast = jsonObj.standard * 10
-    fees.fastest *= 10
-  }
+  const fees = asEvmGasStation(currencyInfo.pluginId, jsonObj)
 
   // Sanity checks
   if (fees.safeLow <= 0 || fees.safeLow > GAS_PRICE_SANITY_CHECK) {