+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: (Polygon) Update Polygon gas station url
+
 ## 4.42.2 (2025-03-25)
 
 - added: (ADA) Add `isSpendable` to `EdgeTransaction.otherParams`.